Update Flutter Renderer rendering logic
Now FlutterActors that extend ActorBasePath will use a FlutterPathMixin to include in their interface the `updatePath()` function
